# docs/openapi.yaml
{"openapi":"3.0.1","info":{"title":"Bank Card Management API","description":"API для управления банковскими картами: аутентификация, CRUD карт, переводы между своими картами","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local Development Server"}],"paths":{"/api/transfers":{"post":{"tags":["transfer-controller"],"summary":"Transfer between own cards","description":"Transfers amount between two cards belonging to the authenticated user. Requires JWT in cookie.","operationId":"transfer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferRequest"}}},"required":true},"responses":{"400":{"description":"Invalid transfer data or insufficient balance","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransferResponse"}}}},"403":{"description":"Forbidden - not USER role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransferResponse"}}}},"200":{"description":"Transfer successful","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransferResponse"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransferResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/admin/users":{"get":{"tags":["admin-user-controller"],"summary":"Get all users","description":"Returns paginated list of all users. Requires JWT in cookie.","operationId":"getAllUsers","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"Users retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUserDTO"}}}},"403":{"description":"Forbidden - not ADMIN role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUserDTO"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUserDTO"}}}}}},"post":{"tags":["admin-user-controller"],"summary":"Create a new user","description":"Creates a new user (USER or ADMIN) by admin. Requires JWT in cookie.","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"201":{"description":"User created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"400":{"description":"Invalid user data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"403":{"description":"Forbidden - not ADMIN role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/admin/cards":{"get":{"tags":["admin-card-controller"],"summary":"Get all cards","description":"Returns paginated list of all cards in the system. Requires JWT in cookie.","operationId":"getAllCards","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Unauthorized - invalid or missing JWT","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCardDTO"}}}},"403":{"description":"Forbidden - user does not have ADMIN role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCardDTO"}}}},"200":{"description":"Cards retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCardDTO"}}}}}},"post":{"tags":["admin-card-controller"],"summary":"Create a new card for a user","description":"Creates a new card for any user by admin. Requires JWT in cookie.","operationId":"createCard","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCardRequest"}}},"required":true},"responses":{"400":{"description":"Invalid card data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDTO"}}}},"403":{"description":"Forbidden - not ADMIN role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDTO"}}}},"201":{"description":"Card created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDTO"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDTO"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDTO"}}}}}}},"/api/admin/cards/{id}/unblock":{"post":{"tags":["admin-card-controller"],"summary":"Unblock a card","description":"Unblocks any card by ID. Requires JWT in cookie.","operationId":"unblockCard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Card not found"},"403":{"description":"Forbidden - not ADMIN role"},"204":{"description":"Card unblocked"},"401":{"description":"Unauthorized"}}}},"/api/admin/cards/{id}/block":{"post":{"tags":["admin-card-controller"],"summary":"Block a card","description":"Blocks any card by ID. Requires JWT in cookie.","operationId":"blockCard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Card not found"},"204":{"description":"Card blocked"},"403":{"description":"Forbidden - not ADMIN role"},"401":{"description":"Unauthorized"}}}},"/api/admin/users/{id}":{"delete":{"tags":["admin-user-controller"],"summary":"Delete a user","description":"Deletes a user by ID. Requires JWT in cookie.","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"User deleted"},"403":{"description":"Forbidden - not ADMIN role"},"404":{"description":"User not found"},"401":{"description":"Unauthorized"}}},"patch":{"tags":["admin-user-controller"],"summary":"Update a user","description":"Updates user details (username, email, role). Requires JWT in cookie.","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"200":{"description":"User updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"400":{"description":"Invalid user data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"403":{"description":"Forbidden - not ADMIN role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/cards/{id}":{"get":{"tags":["user-card-controller"],"summary":"Get card by ID","description":"Returns details of a specific card if it belongs to the user. Requires JWT in cookie.","operationId":"getCardById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Card not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDTO"}}}},"403":{"description":"Forbidden","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDTO"}}}},"200":{"description":"Card found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDTO"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardDTO"}}}}}}},"/api/cards/my":{"get":{"tags":["user-card-controller"],"summary":"Get current user's cards","description":"Returns paginated list of cards belonging to the authenticated user with optional status filter. Requires JWT in cookie.","operationId":"getMyCards","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"401":{"description":"Unauthorized - invalid or missing JWT","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCardDTO"}}}},"403":{"description":"Forbidden - user does not have USER role","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCardDTO"}}}},"200":{"description":"Cards retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageCardDTO"}}}}}}}},"components":{"schemas":{"TransferRequest":{"required":["amount","fromCardId","toCardId"],"type":"object","properties":{"fromCardId":{"type":"string","format":"uuid"},"toCardId":{"type":"string","format":"uuid"},"amount":{"type":"number"}}},"TransferResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"fromCardId":{"type":"string","format":"uuid"},"toCardId":{"type":"string","format":"uuid"},"amount":{"type":"number"},"timestamp":{"type":"string","format":"date-time"},"status":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"message":{"type":"string"}}},"CreateUserRequest":{"required":["email","password","role","username"],"type":"object","properties":{"username":{"maxLength":50,"minLength":3,"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"},"email":{"type":"string"},"role":{"type":"string"}}},"UserDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateCardRequest":{"required":["cardNumber","expiryDate","ownerName"],"type":"object","properties":{"cardNumber":{"maxLength":16,"minLength":16,"type":"string"},"ownerName":{"type":"string"},"expiryDate":{"type":"string"}}},"CardDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"maskedNumber":{"type":"string"},"ownerName":{"type":"string"},"expiryDate":{"type":"string","format":"date"},"status":{"type":"string"},"balance":{"type":"number"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"UpdateUserRequest":{"type":"object","properties":{"username":{"maxLength":50,"minLength":3,"type":"string"},"email":{"type":"string"},"role":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageCardDTO":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/CardDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"PageUserDTO":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}}}}}