# src/main/java/com/example/bankcards/dto/UserDTO.java
package com.example.bankcards.dto;

import java.time.LocalDateTime;
import java.util.UUID;

public record UserDTO(
        UUID id,
        String username,
        String email,
        String role,
        LocalDateTime createdAt,
        LocalDateTime updatedAt
) { }